adding script to start mutt from dmenu + starting mutt by default

+++ resources/configure/default/pre_configure.sh
@@ -1,18 +1,26 @@
 #!/usr/bin/env bash
 
-[ -d $HOME/.password-store ] && rm -rf $HOME/backup_password-store && mv $HOME/.password-store $HOME/backup_password-store
-
-git clone rklopfer@git.klopfer.us:repos/password-store $HOME/.password-store || { 
-	echo "Failed to clone password-store"
-	exit 1
-}
-
-scp git.klopfer.us:.privkey.asc $HOME/.privkey.asc || {
-	echo "Failed to obtain private key"
-	exit 1
-}
+if ! [ -e $HOME/.privkey.asc ]; then
+	# download private key or fail
+	scp git.klopfer.us:.privkey.asc $HOME/.privkey.asc || {
+		echo "Failed to obtain private key"
+		exit 1
+	}
+fi
 
 gpg2 --import $HOME/.privkey.asc || {
 	echo "Failed to import private key"
 	exit 1
 }
+
+if ! [ -d $HOME/.password-store ]; then
+	# clone password repository or fail
+	git clone rklopfer@git.klopfer.us:repos/password-store $HOME/.password-store || { 
+		echo "Failed to clone password-store"
+		exit 1
+	}
+else
+	# update password repository.. or don't fail
+	cd $HOME/.password-store
+	git pull
+fi
